Commit
Update including the bug fix from dwblmo
https://github.com/dwblmo/LMO_PHP8/commit/77fe0317cc9a246f0497b85e956ce1d2f7c14e86

+++ lmo/addon/tipp/lmo-tippcalceinsicht.php
@@ -1,4 +1,4 @@
-<?php
+<?php 
 /** Liga Manager Online 4
   *
   * http://lmo.sourceforge.net/
@@ -7,7 +7,7 @@
   * modify it under the terms of the GNU General Public License as
   * published by the Free Software Foundation; either version 2 of
   * the License, or (at your option) any later version.
-  *
+  * 
   * This program is distributed in the hope that it will be useful,
   * but WITHOUT ANY WARRANTY; without even the implied warranty of
   * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
@@ -16,42 +16,43 @@
   * REMOVING OR CHANGING THE COPYRIGHT NOTICES IS NOT ALLOWED!
   *
   */
-
-if($lmtype != 0) {
+  
+  
+if ($lmtype != 0) {
   $anzsp = $anzteams;
   for($i = 0; $i < $st; $i++) {
     $anzsp = $anzsp/2;
   }
-  if(($klfin == 1) && ($st == $anzst)) {
+  if (($klfin == 1) && ($st == $anzst)) {
     $anzsp = $anzsp+1;
   }
 }
 
 $einsichtfile = PATH_TO_ADDONDIR."/tipp/".$tipp_dirtipp."einsicht/".substr($file, 0, -4)."_".$st.".ein";
-if(!file_exists($einsichtfile)) {
+if (!file_exists($einsichtfile)) {
   echo getMessage($text['tipp'][17],TRUE);
 } else {
   $datei = fopen($einsichtfile, "rb");
   $anztipper = 0;
-  if($datei != false) {
+  if ($datei != false) {
     $tippdaten = array();
     $sekt = "";
     while (!feof($datei)) {
       $zeile = fgets($datei, 10000);
       $zeile = trim($zeile);
-      if((substr($zeile, 0, 1) == "@") && (substr($zeile, -1) == "@")) {
+      if ((substr($zeile, 0, 1) == "@") && (substr($zeile, -1) == "@")) {
         $jkwert = trim(substr($zeile, 1, -1));
       } elseif((substr($zeile, 0, 1) == "[") && (substr($zeile, -1) == "]")) {
         $sekt = trim(substr($zeile, 1, -1));
         $jkwert = -1;
-        if($zeile != "[Options]") {
+        if ($zeile != "[Options]") {
           array_push($tippdaten, $sekt."|||EOL");
           $anztipper++;
         }
       } elseif((strpos($zeile, "=") != false) && (substr($zeile, 0, 1) != ";")) {
         $schl = trim(substr($zeile, 0, strpos($zeile, "=")));
         $wert = trim(substr($zeile, strpos($zeile, "=")+1));
-        if(!isset($jkwert)) {
+        if (!isset($jkwert)) {
           $jkwert = -1;
         }
         array_push($tippdaten, $sekt."|".$schl."|".$wert."|".$jkwert."|EOL");
@@ -59,13 +60,13 @@ if(!file_exists($einsichtfile)) {
     }
     fclose($datei);
   }
-
+   
   $tippernick = array_pad($array, $anztipper+1, "");
   $jksp2 = array_pad($array, $anztipper+1, "");
   $tippa = array_pad($array, $anztipper+1, "");
   $tippb = array_pad($array, $anztipper+1, "");
   for($i = 0; $i < $anztipper; $i++) {
-    if($lmtype != 0) {
+    if ($lmtype != 0) {
       $tippa[$i] = array_pad($array, $anzsp+1, "");
       $tippb[$i] = array_pad($array, $anzsp+1, "");
       for($n = 0; $n < $anzsp; $n++) {
@@ -77,8 +78,8 @@ if(!file_exists($einsichtfile)) {
       $tippb[$i] = array_pad(array("_"), $anzsp+1, "_");
     }
   }
-
-  if($lmtype == 0) {
+   
+  if ($lmtype == 0) {
     $tendenz1 = array_pad(array("0"), $anzsp+1, "0");
     $tendenz0 = array_pad(array("0"), $anzsp+1, "0");
     $tendenz2 = array_pad(array("0"), $anzsp+1, "0");
@@ -104,51 +105,51 @@ if(!file_exists($einsichtfile)) {
       $btip[$i] = array_pad(array("false"), $modus[$st-1]+1, "false");
     }
   }
-
+   
   $t = 0;
-  if($endtab < 1) {
+  if ($endtab < 1) {
     $endtab = $anzst;
   }
   for($i = 1; $i <= count($tippdaten); $i++) {
     $dum = explode('|', $tippdaten[$i-1]);
     $op1 = $dum[0];
     // Nick
-    $op3 = (int)substr($dum[1], 2)-1;
+    $op3 = substr($dum[1], 2)-1;
     // Spieltagsnummer
     $op4 = substr($dum[1], 0, 2);
     // TP
     $op6 = (int)substr($dum[1], 2, -1)-1;
     $op7 = (int)substr($dum[1], -1)-1;
     $op8 = $dum[3];
-    if($tippernick[$t] != $op1) {
-      if($tippernick[$t] != "") {
+    if ($tippernick[$t] != $op1) {
+      if ($tippernick[$t] != "") {
         $t++;
       }
       $tippernick[$t] = $op1;
     }
     $jksp2[$t] = $op8;
-    if($op4 == "GA") {
-      if($lmtype != 0) {
+    if ($op4 == "GA") {
+      if ($lmtype != 0) {
         $tippa[$t][$op6][$op7] = $dum[2];
-        if($dum[2] > 0) {
+        if ($dum[2] > 0) {
           $toregesa[$op6][$op7] += $dum[2];
         }
       } else {
         $tippa[$t][$op3] = $dum[2];
-        if($dum[2] > 0) {
+        if ($dum[2] > 0) {
           $toregesa[$op3] += $dum[2];
         }
       }
     }
-    if($op4 == "GB") {
-      if($lmtype != 0) {
+    if ($op4 == "GB") {
+      if ($lmtype != 0) {
         $tippb[$t][$op6][$op7] = $dum[2];
-        if($dum[2] >= 0) {
+        if ($dum[2] >= 0) {
           $toregesb[$op6][$op7] += $dum[2];
           $anzgetippt[$op6][$op7]++;
         }
-
-        if($tippb[$t][$op6][$op7] < $tippa[$t][$op6][$op7]) {
+         
+        if ($tippb[$t][$op6][$op7] < $tippa[$t][$op6][$op7]) {
           $tendenz1[$op6][$op7]++;
         } elseif($tippb[$t][$op6][$op7] > $tippa[$t][$op6][$op7]) {
           $tendenz2[$op6][$op7]++;
@@ -157,12 +158,12 @@ if(!file_exists($einsichtfile)) {
         }
       } else {
         $tippb[$t][$op3] = $dum[2];
-        if($dum[2] >= 0) {
+        if ($dum[2] >= 0) {
           $toregesb[$op3] += $dum[2];
           $anzgetippt[$op3]++;
         }
-
-        if($tippb[$t][$op3] < $tippa[$t][$op3]) {
+         
+        if ($tippb[$t][$op3] < $tippa[$t][$op3]) {
           $tendenz1[$op3]++;
         } elseif($tippb[$t][$op3] > $tippa[$t][$op3]) {
           $tendenz2[$op3]++;
@@ -172,8 +173,8 @@ if(!file_exists($einsichtfile)) {
       }
     }
   }
-
-  if($todo = "einsicht") {
+   
+  if ($todo = "einsicht") {
     $tab0 = array();
     for($a = 0; $a < $anztipper; $a++) {
       array_push($tab0, strtolower($tippernick[$a]).(50000000+$a));
